2.7.0b - Including tarball for Unity XR package. Lots of bug fixes.
Changes for 2.7.0

 * Updated sdk header to 1.14.15

 * Moved OpenVR Unity XR package to a tarball included in this project instead of NPM per Unity's new TOS.

 * Removing app key from binding files on build (fixes some scenarios where an indev binding file wouldn't apply to a released app)

 * Retargeted some DLLs to not cause issues on other platforms

 * Fixed an issue with some builds where the settings file wouldn't copy over properly.

 * Removed some resolution dialog recommendations for 2019+

 * Reworked the auto-enable-vr systems to be easier to manage in the future and cause less problems in the present.

 * Added initialPriority to SteamVR_ActivateActionSetOnLoad. Thanks for the pull request @shiena

 * Fixed ControllerButtonHints for URP, Thanks again for the pull request @shiena

 * Added RemoveAllListeners to each action type for easy delegate disposal. Thanks for the pull request @Extrys

 * Unity XR: Fix for laptops not always using the correct video card

 * Unity XR: Fixing IL2CPP issue with builds

 * Unity XR: Enabled user presence detection via the userPresence feature.

 * Unity XR: Temporary fix for Unity XR Settings reverting to defaults when users hit play in the editor and have the window open.
