+++ Assets/SteamVR/Editor/SteamVR_AutoEnableVR_2018to2019.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b894f6f57c99a1c46957650bc11d7824
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
