+++ Assets/SteamVR/Editor/SteamVR_AutoEnableVR_2019plus.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 524ac57f667bffd459f44076a7111efb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
