+++ Assets/SteamVR/Editor/SteamVR_AutoEnableVR_5.4to2018.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f3c98d5dbcd01ee4ba302d8f71a8d55a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
