+++ Assets/SteamVR/Input/Editor/SteamVR_Input_EditorWindow.cs
@@ -293,6 +293,7 @@ namespace Valve.VR
                 SteamVR_CopyExampleInputFiles.CopyFiles(true);
                 System.Threading.Thread.Sleep(1000);
                 bool initializeSuccess = SteamVR_Input.InitializeFile();
+                EditorApplication.delayCall += SaveFile; 
                 return initializeSuccess;
             }
             else
@@ -683,9 +684,12 @@ namespace Valve.VR
             EditorGUILayout.EndFadeGroup();
         }
 
+        private Vector2 setScrollPosition;
         private void DrawSets()
         {
             EditorGUILayout.LabelField("Action Sets", headerLabelStyle);
+
+            setScrollPosition = EditorGUILayout.BeginScrollView(setScrollPosition);
             EditorGUILayout.BeginHorizontal();
 
             GUILayout.FlexibleSpace();
@@ -796,6 +800,8 @@ namespace Valve.VR
             GUILayout.FlexibleSpace();
 
             EditorGUILayout.EndHorizontal();
+
+            EditorGUILayout.EndScrollView();
         }
 
         private static MemberInfo GetMemberInfo<TModel, TItem>(TModel model, Expression<Func<TModel, TItem>> expr)
