+++ Assets/SteamVR/Input/SteamVR_ActivateActionSetOnLoad.cs
@@ -19,13 +19,14 @@ namespace Valve.VR
         public bool activateOnStart = true;
         public bool deactivateOnDestroy = true;
 
+        public int initialPriority = 0;
 
         private void Start()
         {
             if (actionSet != null && activateOnStart)
             {
                 //Debug.Log(string.Format("[SteamVR] Activating {0} action set.", actionSet.fullPath));
-                actionSet.Activate(forSources, 0, disableAllOtherActionSets);
+                actionSet.Activate(forSources, initialPriority, disableAllOtherActionSets);
             }
         }
