--- Assets/SteamVR/Input/SteamVR_DefaultAction.cs
@@ -1,1 +1,0 @@
-﻿
