--- Assets/SteamVR/Input/SteamVR_DefaultActionSet.cs
@@ -1,1 +1,0 @@
-﻿
