--- Assets/SteamVR/OpenVRAutoUpdater/Editor.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: b76f4d0efe922444e8d79781fe8e28bc
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
