--- Assets/SteamVR/OpenVRAutoUpdater/Editor/OpenVRAutoUpdater.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 538300977557ee34892368a50c97bdd6
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
