--- Assets/SteamVR/OpenVRAutoUpdater/Editor/OpenVRSimpleJSON.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ca02308d9febaa442a742ecaaaa06d2c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
