+++ Assets/SteamVR/OpenVRUnityXRPackage.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 6cd256751f738d645b7b64c556da3759
+folderAsset: yes
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
