+++ Assets/SteamVR/OpenVRUnityXRPackage/Editor.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ffd65d6a9f4d5994e8fa33d7d36c5828
+folderAsset: yes
+timeCreated: 1605580376
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
