+++ Assets/SteamVR/OpenVRUnityXRPackage/Editor/OpenVRPackageInstaller.cs.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 538300977557ee34892368a50c97bdd6
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
