+++ Assets/SteamVR/OpenVRUnityXRPackage/Editor/com.valvesoftware.unity.openvr-1.1.1.tgz.meta
@@ -1,0 +1,6 @@
+fileFormatVersion: 2
+guid: 5240c862c299e734e89c74efce519a1a
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
