+++ Assets/SteamVR/readme.txt
@@ -1,4 +1,4 @@
-﻿# SteamVR Unity Plugin - v2.6.1 (sdk 1.13.10)
+﻿# SteamVR Unity Plugin - v2.7.0b (sdk 1.14.15)
 
 Copyright (c) Valve Corporation, All rights reserved.
 
@@ -36,6 +36,37 @@ Input and Steam:
  If you publish your game to steam you can let users change their input bindings while the game is not running by setting the location of your action manifest. On the steamworks partner site go to the Application settings, and the Virtual Reality Section. At the bottom you'll see a radio button to designate your title as a SteamVR Input application. You then can set the location of your action manifest. In older versions of the plugin this was next to the executable. In versions 2.3.3 and above this is in [GameName]_Data/StreamingAssets/SteamVR/actions.json.
 
  
+Changes for 2.7.0
+
+ * Updated sdk header to 1.14.15
+
+ * Moved OpenVR Unity XR package to a tarball included in this project instead of NPM per Unity's new TOS.
+
+ * Removing app key from binding files on build (fixes some scenarios where an indev binding file wouldn't apply to a released app)
+
+ * Retargeted some DLLs to not cause issues on other platforms
+
+ * Fixed an issue with some builds where the settings file wouldn't copy over properly.
+
+ * Removed some resolution dialog recommendations for 2019+
+
+ * Reworked the auto-enable-vr systems to be easier to manage in the future and cause less problems in the present.
+
+ * Added initialPriority to SteamVR_ActivateActionSetOnLoad. Thanks for the pull request @shiena
+
+ * Fixed ControllerButtonHints for URP, Thanks again for the pull request @shiena
+
+ * Added RemoveAllListeners to each action type for easy delegate disposal. Thanks for the pull request @Extrys
+
+ * Unity XR: Fix for laptops not always using the correct video card
+
+ * Unity XR: Fixing IL2CPP issue with builds
+
+ * Unity XR: Enabled user presence detection via the userPresence feature.
+
+ * Unity XR: Temporary fix for Unity XR Settings reverting to defaults when users hit play in the editor and have the window open.
+
+
  
 Changes for 2.6.1
