+++ ProjectSettings/EditorBuildSettings.asset
@@ -7,3 +7,5 @@ EditorBuildSettings:
   m_Scenes:
   - enabled: 1
     path: Assets/SteamVR/InteractionSystem/Samples/Interactions_Example.unity
+  - enabled: 0
+    path: Assets/SteamVR/InteractionSystem/Samples/Interactions_Example copy.unity
