+++ ProjectSettings/ProjectVersion.txt
@@ -1,2 +1,2 @@
-m_EditorVersion: 5.4.0f1
+m_EditorVersion: 5.4.6f3
 m_StandardAssetsVersion: 0
