+++ ProjectSettings/TagManager.asset
@@ -13,7 +13,7 @@ TagManager:
   - UI
   - 
   - 
-  - 
+  - Hand
   - 
   - 
   -
