+++ ProjectSettings/UnityConnectSettings.asset
@@ -7,9 +7,11 @@ UnityConnectSettings:
   m_TestMode: 0
   m_TestEventUrl: 
   m_TestConfigUrl: 
+  m_TestInitMode: 0
   CrashReportingSettings:
     m_EventUrl: https://perf-events.cloud.unity3d.com/api/events/crashes
     m_Enabled: 0
+    m_CaptureEditorExceptions: 1
   UnityPurchasingSettings:
     m_Enabled: 0
     m_TestMode: 0
@@ -19,3 +21,12 @@ UnityConnectSettings:
     m_TestMode: 0
     m_TestEventUrl: 
     m_TestConfigUrl: 
+  UnityAdsSettings:
+    m_Enabled: 0
+    m_InitializeOnStartup: 1
+    m_TestMode: 0
+    m_EnabledPlatforms: 4294967295
+    m_IosGameId: 
+    m_AndroidGameId: 
+  PerformanceReportingSettings:
+    m_Enabled: 0
